Updated the user management and content moderation components for prisma schema
